Revert "Add discovery strategy tracking, observatory view, and mobile shell/valuation chart fixes"
This reverts commit 6167492c4d75ce7060cad30b5d14d7a336dae9c2.
