+++ frontend/src/app/(app)/dashboard/[ticker]/summary/page.tsx
@@ -65,11 +65,11 @@ function fmtDateTimeNoSeconds(value: string): string {
   });
 }
 
-function fmtMoney(value: number | null | undefined, currency: string = "USD"): string {
+function fmtMoney(value: number | null | undefined): string {
   if (typeof value !== "number" || !Number.isFinite(value)) return "N/A";
   return new Intl.NumberFormat(undefined, {
     style: "currency",
-    currency,
+    currency: "USD",
     minimumFractionDigits: 2,
     maximumFractionDigits: 2,
   }).format(value);
@@ -118,12 +118,10 @@ function MeanTable({
   title,
   rows,
   liveCurrentPrice,
-  currencyCode,
 }: {
   title: string;
   rows: MeanRow[];
   liveCurrentPrice: number | null;
-  currencyCode: string;
 }) {
   return (
     <section className="rounded-2xl border border-white/10 bg-zinc-950/70 p-4">
@@ -135,7 +133,7 @@ function MeanTable({
             <article key={`${row.key}-mobile`} className="rounded-xl border border-white/10 bg-black/30 p-3">
               <p className="text-[11px] uppercase tracking-[0.12em] text-zinc-500">{row.label}</p>
               <p className={`mt-1 text-xl font-bold ${toneClassFromTarget(row.mean_target_price, liveCurrentPrice)}`}>
-                {fmtMoney(row.mean_target_price, currencyCode)}
+                {fmtMoney(row.mean_target_price)}
               </p>
               <p className={`mt-1 text-xs font-semibold ${toneClassFromSign(changePct)}`}>
                 Change vs Live: ({fmtPct(changePct)})
@@ -156,35 +154,28 @@ function MeanTable({
             <tr>
               <th className="px-3 py-2 text-left font-medium">Name</th>
               <th className="px-3 py-2 text-right font-medium">Mean Target</th>
-              <th className="px-3 py-2 text-right font-medium">Change vs Live</th>
               <th className="px-3 py-2 text-right font-medium">Target N</th>
               <th className="px-3 py-2 text-right font-medium">Mean Allocation</th>
               <th className="px-3 py-2 text-right font-medium">Allocation N</th>
             </tr>
           </thead>
           <tbody>
-            {rows.map((row) => {
-              const changePct = targetChangePct(row.mean_target_price, liveCurrentPrice);
-              return (
-                <tr key={row.key} className="border-b border-white/5 last:border-b-0">
-                  <td className="px-3 py-2 font-medium text-zinc-200">{row.label}</td>
-                  <td className={`px-3 py-2 text-right ${toneClassFromTarget(row.mean_target_price, liveCurrentPrice)}`}>
-                    {fmtMoney(row.mean_target_price, currencyCode)}
-                  </td>
-                  <td className={`px-3 py-2 text-right ${toneClassFromSign(changePct)}`}>
-                    {fmtPct(changePct)}
-                  </td>
-                  <td className="px-3 py-2 text-right text-zinc-400">{row.target_samples}</td>
-                  <td className={`px-3 py-2 text-right ${toneClassFromSign(row.mean_allocation_pct)}`}>
-                    {fmtPct(row.mean_allocation_pct)}
-                  </td>
-                  <td className="px-3 py-2 text-right text-zinc-400">{row.allocation_samples}</td>
-                </tr>
-              );
-            })}
+            {rows.map((row) => (
+              <tr key={row.key} className="border-b border-white/5 last:border-b-0">
+                <td className="px-3 py-2 font-medium text-zinc-200">{row.label}</td>
+                <td className={`px-3 py-2 text-right ${toneClassFromTarget(row.mean_target_price, liveCurrentPrice)}`}>
+                  {fmtMoney(row.mean_target_price)}
+                </td>
+                <td className="px-3 py-2 text-right text-zinc-400">{row.target_samples}</td>
+                <td className={`px-3 py-2 text-right ${toneClassFromSign(row.mean_allocation_pct)}`}>
+                  {fmtPct(row.mean_allocation_pct)}
+                </td>
+                <td className="px-3 py-2 text-right text-zinc-400">{row.allocation_samples}</td>
+              </tr>
+            ))}
             {!rows.length ? (
               <tr>
-                <td colSpan={6} className="px-3 py-3 text-zinc-500">
+                <td colSpan={5} className="px-3 py-3 text-zinc-500">
                   No rows available.
                 </td>
               </tr>
@@ -255,7 +246,6 @@ export default function DashboardSummaryPage({
   const [data, setData] = useState<SummaryPayload | null>(null);
 
   const reportId = search?.get("report") || "";
-  const currencyCode = upper.endsWith(".TA") ? "ILS" : "USD";
 
   useEffect(() => {
     let cancelled = false;
@@ -361,12 +351,12 @@ export default function DashboardSummaryPage({
           <section className="grid gap-4 md:grid-cols-4">
             <article className="rounded-2xl border border-white/10 bg-zinc-950/70 p-4">
               <p className="text-xs uppercase tracking-[0.18em] text-zinc-500">Current Live Price</p>
-              <p className="mt-2 text-3xl font-bold text-zinc-100">{fmtMoney(data.overview.live_current_price, currencyCode)}</p>
+              <p className="mt-2 text-3xl font-bold text-zinc-100">{fmtMoney(data.overview.live_current_price)}</p>
             </article>
             <article className="rounded-2xl border border-white/10 bg-zinc-950/70 p-4">
               <p className="text-xs uppercase tracking-[0.18em] text-zinc-500">Mean Target</p>
               <p className={`mt-2 text-3xl font-bold ${toneClassFromTarget(data.overview.mean_target_price, data.overview.live_current_price)}`}>
-                {fmtMoney(data.overview.mean_target_price, currencyCode)}{" "}
+                {fmtMoney(data.overview.mean_target_price)}{" "}
                 <span className="text-lg">({fmtPct(meanTargetChangePct)})</span>
               </p>
               <p className="mt-2 text-xs text-zinc-400">N {data.overview.target_samples}</p>
@@ -389,13 +379,11 @@ export default function DashboardSummaryPage({
             title="By Model Mean Target + Mean Allocation"
             rows={modelRowsSorted}
             liveCurrentPrice={data.overview.live_current_price}
-            currencyCode={currencyCode}
           />
           <MeanTable
             title="By Valuator Mean Target + Mean Allocation"
             rows={valuatorRowsSorted}
             liveCurrentPrice={data.overview.live_current_price}
-            currencyCode={currencyCode}
           />
           <AssumptionsTable rows={data.assumptions} />
         </>
