+++ frontend/src/app/(app)/discovery/page.tsx
@@ -1,15 +1,13 @@
 "use client";
 
 import Link from "next/link";
-import { useEffect, useMemo, useState } from "react";
+import { useEffect, useState } from "react";
 import type { ReactNode } from "react";
 import { Gem, Radar, ShieldAlert, TrendingDown, TrendingUp } from "lucide-react";
-import { Line, LineChart, ResponsiveContainer, Tooltip, XAxis, YAxis } from "recharts";
 
 import type { DiscoveryRow } from "@/lib/dashboard-types";
 
 type DiscoveryLensType = "overall" | "model" | "valuator";
-type PerformanceWindowKey = "1w" | "1m" | "1y" | "all";
 
 type DiscoveryPayload = {
   generated_at: string;
@@ -31,65 +29,6 @@ type DiscoveryPayload = {
   top_lowest_allocation: DiscoveryRow[];
 };
 
-type DiscoveryPerformancePoint = {
-  date: string;
-  nav: number;
-  cumulative_return_pct: number;
-};
-
-type DiscoveryPerformanceSeries = {
-  key: string;
-  label: string;
-  points: DiscoveryPerformancePoint[];
-  latest_stats: {
-    nav: number | null;
-    cumulative_return_pct: number | null;
-    daily_return_pct: number | null;
-    max_drawdown_pct: number | null;
-  };
-};
-
-type DiscoveryPerformancePayload = {
-  generated_at: string;
-  trade_date: string | null;
-  lens: {
-    type: DiscoveryLensType;
-    key: string | null;
-    label: string;
-  };
-  series: DiscoveryPerformanceSeries[];
-  windows: Record<string, { key: string; label: string; start_date: string | null }>;
-};
-
-type ChartRow = {
-  date: string;
-  [key: string]: number | string | null;
-};
-
-async function readJsonOrThrow<T>(res: Response, label: string): Promise<T> {
-  const raw = await res.text();
-  if (!res.ok) {
-    const details = raw.trim() ? `: ${raw.slice(0, 240)}` : "";
-    throw new Error(`${label} failed (${res.status})${details}`);
-  }
-  if (!raw.trim()) {
-    throw new Error(`${label} returned an empty response body`);
-  }
-  try {
-    return JSON.parse(raw) as T;
-  } catch {
-    throw new Error(`${label} returned invalid JSON`);
-  }
-}
-
-const LINE_COLORS: Record<string, string> = {
-  most_undervalued_top10: "#2563eb",
-  most_undervalued_top20: "#7c3aed",
-  highest_allocation_top10: "#0ea5e9",
-  highest_allocation_top20: "#f97316",
-  benchmark_sp500: "#facc15",
-};
-
 function fmtDateTimeNoSeconds(value: string): string {
   const dt = new Date(value);
   if (!Number.isFinite(dt.getTime())) return "N/A";
@@ -107,72 +46,10 @@ function fmtPct(value: number): string {
   return `${value >= 0 ? "+" : ""}${value.toFixed(2)}%`;
 }
 
-function returnToneClass(value: number | null | undefined): string {
-  if (typeof value !== "number" || !Number.isFinite(value)) return "text-zinc-300";
-  if (Math.abs(value) < 1e-9) return "text-zinc-200";
-  return value > 0 ? "text-emerald-300" : "text-red-300";
-}
-
 function decisionClass(tone?: DiscoveryRow["decision_tone"]): string {
   if (tone === "buy") return "hib-signal-buy";
   if (tone === "sell") return "hib-signal-sell";
   return "hib-signal-hold";
-}
-
-function buildWindowSeries(
-  series: DiscoveryPerformanceSeries[],
-  startDate: string | null,
-): DiscoveryPerformanceSeries[] {
-  return series.map((line) => {
-    const sliced = startDate
-      ? line.points.filter((point) => point.date >= startDate)
-      : line.points.slice();
-    if (!sliced.length) {
-      return {
-        ...line,
-        points: [],
-        latest_stats: {
-          nav: null,
-          cumulative_return_pct: null,
-          daily_return_pct: line.latest_stats.daily_return_pct,
-          max_drawdown_pct: null,
-        },
-      };
-    }
-    const baseline = sliced[0].nav;
-    const points = sliced.map((point) => ({
-      ...point,
-      cumulative_return_pct: baseline > 0 ? ((point.nav / baseline) - 1) * 100 : 0,
-    }));
-    const last = points[points.length - 1];
-    return {
-      ...line,
-      points,
-      latest_stats: {
-        ...line.latest_stats,
-        nav: last.nav,
-        cumulative_return_pct: last.cumulative_return_pct,
-      },
-    };
-  });
-}
-
-function buildChartRows(series: DiscoveryPerformanceSeries[]): ChartRow[] {
-  const allDates = new Set<string>();
-  for (const line of series) {
-    for (const point of line.points) {
-      allDates.add(point.date);
-    }
-  }
-  const dates = Array.from(allDates).sort((a, b) => a.localeCompare(b));
-  return dates.map((date) => {
-    const row: ChartRow = { date };
-    for (const line of series) {
-      const point = line.points.find((p) => p.date === date) || null;
-      row[line.key] = point ? point.cumulative_return_pct : null;
-    }
-    return row;
-  });
 }
 
 function SectionCard({
@@ -275,38 +152,24 @@ function SectionCard({
 
 export default function DiscoveryPage() {
   const [data, setData] = useState<DiscoveryPayload | null>(null);
-  const [performance, setPerformance] = useState<DiscoveryPerformancePayload | null>(null);
-  const [loadError, setLoadError] = useState<string | null>(null);
   const [loading, setLoading] = useState(true);
   const [lensType, setLensType] = useState<DiscoveryLensType>("overall");
   const [lensKey, setLensKey] = useState("");
-  const [windowKey, setWindowKey] = useState<PerformanceWindowKey>("1m");
 
   useEffect(() => {
     let cancelled = false;
     async function run() {
       setLoading(true);
-      setLoadError(null);
       try {
         const params = new URLSearchParams();
         params.set("lens_type", lensType);
         if (lensType !== "overall" && lensKey.trim()) {
           params.set("lens_key", lensKey.trim());
         }
-        const [discoveryRes, performanceRes] = await Promise.all([
-          fetch(`/api/discovery?${params.toString()}`, { cache: "no-store" }),
-          fetch(`/api/discovery/performance?${params.toString()}`, { cache: "no-store" }),
-        ]);
-        const discoveryJson = await readJsonOrThrow<DiscoveryPayload>(discoveryRes, "Discovery API");
-        const performanceJson = await readJsonOrThrow<DiscoveryPerformancePayload>(performanceRes, "Discovery performance API");
+        const res = await fetch(`/api/discovery?${params.toString()}`, { cache: "no-store" });
+        const json = (await res.json()) as DiscoveryPayload;
         if (!cancelled) {
-          setData(discoveryJson);
-          setPerformance(performanceJson);
-        }
-      } catch (err) {
-        if (!cancelled) {
-          setLoadError(err instanceof Error ? err.message : "Failed loading discovery data");
-          setPerformance(null);
+          setData(json);
         }
       } finally {
         if (!cancelled) {
@@ -334,18 +197,6 @@ export default function DiscoveryPage() {
     setLensKey(pool[0] || "");
   };
 
-  const chartWindowStart = useMemo(() => {
-    if (!performance) return null;
-    return performance.windows?.[windowKey]?.start_date || null;
-  }, [performance, windowKey]);
-
-  const windowedSeries = useMemo(
-    () => buildWindowSeries(performance?.series || [], chartWindowStart),
-    [performance, chartWindowStart],
-  );
-
-  const chartRows = useMemo(() => buildChartRows(windowedSeries), [windowedSeries]);
-
   return (
     <div className="mx-auto w-full max-w-[1400px] px-4 py-6 text-zinc-100 sm:px-8">
       <div>
@@ -354,23 +205,14 @@ export default function DiscoveryPage() {
           <p className="text-xs uppercase tracking-[0.2em] text-zinc-500">All Reports (Latest Per Ticker)</p>
         </header>
 
-        {loading ? (
+        {loading || !data ? (
           <div className="grid gap-4 md:grid-cols-3">
             {Array.from({ length: 6 }).map((_, idx) => (
               <div key={idx} className="h-40 animate-pulse rounded-xl border border-white/10 bg-white/5" />
             ))}
           </div>
-        ) : !data ? (
-          <div className="rounded-xl border border-red-400/30 bg-red-500/10 px-4 py-3 text-sm text-red-100">
-            {loadError || "Failed to load discovery data."}
-          </div>
         ) : (
           <>
-            {loadError ? (
-              <div className="mb-4 rounded-xl border border-amber-400/30 bg-amber-500/10 px-3 py-2 text-sm text-amber-100">
-                {loadError}
-              </div>
-            ) : null}
             <section className="mb-4 rounded-2xl border border-white/10 bg-zinc-950/70 p-4">
               <div className="flex flex-col gap-3 sm:flex-row sm:items-center sm:justify-between">
                 <div>
@@ -424,77 +266,6 @@ export default function DiscoveryPage() {
                   ) : null}
                 </div>
               </div>
-            </section>
-            <section className="mb-4 rounded-2xl border border-white/10 bg-zinc-950/70 p-4">
-              <div className="mb-3 flex flex-col gap-3 sm:flex-row sm:items-center sm:justify-between">
-                <div>
-                  <p className="text-xs uppercase tracking-[0.14em] text-zinc-500">Strategy Equity Curves</p>
-                  <p className="text-sm text-zinc-300">Dynamic daily-rebalanced forward test vs S&P 500.</p>
-                </div>
-                <div className="inline-flex rounded-xl border border-white/15 bg-white/5 p-1 text-xs uppercase tracking-[0.12em]">
-                  {(["1w", "1m", "1y", "all"] as PerformanceWindowKey[]).map((key) => (
-                    <button
-                      key={key}
-                      type="button"
-                      onClick={() => setWindowKey(key)}
-                      className={`rounded-lg px-3 py-1.5 transition ${
-                        windowKey === key ? "bg-emerald-500/20 text-emerald-100" : "text-zinc-300 hover:text-zinc-100"
-                      }`}
-                    >
-                      {performance?.windows?.[key]?.label || key.toUpperCase()}
-                    </button>
-                  ))}
-                </div>
-              </div>
-
-              {chartRows.length ? (
-                <>
-                  <div className="h-72 w-full">
-                    <ResponsiveContainer width="100%" height="100%">
-                      <LineChart data={chartRows} margin={{ left: 12, right: 12, top: 8, bottom: 8 }}>
-                        <XAxis dataKey="date" tick={{ fontSize: 11 }} minTickGap={28} />
-                        <YAxis tick={{ fontSize: 11 }} tickFormatter={(v) => `${Number(v).toFixed(0)}%`} width={56} />
-                        <Tooltip
-                          formatter={(value) => `${Number(value ?? 0).toFixed(2)}%`}
-                          labelFormatter={(value) => String(value)}
-                        />
-                        {windowedSeries.map((line) => (
-                          <Line
-                            key={line.key}
-                            type="monotone"
-                            dataKey={line.key}
-                            stroke={LINE_COLORS[line.key] || "#a1a1aa"}
-                            strokeWidth={line.key === "benchmark_sp500" ? 2.5 : 2}
-                            dot={false}
-                            connectNulls
-                          />
-                        ))}
-                      </LineChart>
-                    </ResponsiveContainer>
-                  </div>
-                  <div className="mt-3 grid gap-2 sm:grid-cols-2 lg:grid-cols-3 xl:grid-cols-5">
-                    {windowedSeries.map((line) => (
-                      <div key={`stat-${line.key}`} className="rounded-lg border border-white/10 bg-black/25 p-2 text-xs">
-                        <p className="flex items-center gap-1.5 text-zinc-400">
-                          <span
-                            className="inline-block h-2 w-2 rounded-full"
-                            style={{ backgroundColor: LINE_COLORS[line.key] || "#a1a1aa" }}
-                            aria-hidden
-                          />
-                          <span>{line.label}</span>
-                        </p>
-                        <p className={`mt-1 font-semibold ${returnToneClass(line.latest_stats.cumulative_return_pct)}`}>
-                          {typeof line.latest_stats.cumulative_return_pct === "number"
-                            ? fmtPct(line.latest_stats.cumulative_return_pct)
-                            : "N/A"}
-                        </p>
-                      </div>
-                    ))}
-                  </div>
-                </>
-              ) : (
-                <p className="text-sm text-zinc-500">Performance history is not available yet. It will start at feature go-live.</p>
-              )}
             </section>
             <p className="mb-4 text-sm text-zinc-400">
               Scanned {data.count} tickers for this lens. Generated at {fmtDateTimeNoSeconds(String(data.generated_at || ""))}.
