+++ frontend/src/components/target-price-chart.tsx
@@ -19,12 +19,6 @@ import { buildCurrencyContext, fmtMoney, type CurrencyContext } from "@/componen
 import { useThemeTokens } from "@/lib/theme-tokens";
 
 const CHART_TOKENS = ["--chart-grid", "--chart-current", "--chart-bull", "--chart-bear"] as const;
-const CHART_FALLBACKS: Record<(typeof CHART_TOKENS)[number], string> = {
-  "--chart-grid": "#cbd5e1",
-  "--chart-current": "#f59e0b",
-  "--chart-bull": "#16a34a",
-  "--chart-bear": "#dc2626",
-};
 
 type ChartHoverState = {
   chartX?: number;
@@ -56,14 +50,6 @@ function ChartHoverTooltip({
 export function TargetPriceChart({ data }: { data: DashboardPayload | null }) {
   const currencyContext = useMemo(() => buildCurrencyContext(data), [data]);
   const tokens = useThemeTokens(CHART_TOKENS);
-  const resolvedTokens = useMemo(() => {
-    const out = {} as Record<(typeof CHART_TOKENS)[number], string>;
-    for (const key of CHART_TOKENS) {
-      const value = String(tokens[key] || "").trim();
-      out[key] = value || CHART_FALLBACKS[key];
-    }
-    return out;
-  }, [tokens]);
   const consensus = data?.valuation_hub?.consensus;
   const consensusCurrent =
     typeof consensus?.current_price === "number" && Number.isFinite(consensus.current_price)
@@ -193,7 +179,7 @@ export function TargetPriceChart({ data }: { data: DashboardPayload | null }) {
       <div ref={wrapRef} className="hib-chart relative h-96">
         <ResponsiveContainer width="100%" height="100%">
           <BarChart data={chartData} onMouseMove={handleMouseMove} onMouseLeave={hide}>
-            <CartesianGrid strokeDasharray="3 3" stroke={resolvedTokens["--chart-grid"]} />
+            <CartesianGrid strokeDasharray="3 3" stroke={tokens["--chart-grid"]} />
             <XAxis dataKey="name" tick={false} axisLine={false} tickLine={false} />
             <YAxis
               width={140}
@@ -214,7 +200,7 @@ export function TargetPriceChart({ data }: { data: DashboardPayload | null }) {
             {Number(consensus?.current_price || 0) > 0 ? (
               <ReferenceLine
                 y={Number(consensus?.current_price || 0)}
-                stroke={resolvedTokens["--chart-current"]}
+                stroke={tokens["--chart-current"]}
                 strokeWidth={2.5}
                 strokeDasharray="6 4"
               />
@@ -223,7 +209,7 @@ export function TargetPriceChart({ data }: { data: DashboardPayload | null }) {
               {chartData.map((entry) => (
                 <Cell
                   key={`target-${entry.name}`}
-                  fill={entry.aboveCurrent ? resolvedTokens["--chart-bull"] : resolvedTokens["--chart-bear"]}
+                  fill={entry.aboveCurrent ? tokens["--chart-bull"] : tokens["--chart-bear"]}
                   style={{ cursor: "pointer" }}
                 />
               ))}
