+++ frontend/src/lib/hit-rate-aggregate.ts
@@ -1,5 +1,4 @@
 import type { DashboardPayload } from "./dashboard-types";
-import { canonicalMethodName } from "./method-names";
 import {
   actualDirectionFromPrices,
   allocationDirectionFromAmount,
@@ -138,12 +137,12 @@ export function computeHitRateAggregation(
     const modelRows =
       methodTabs.length > 0
         ? methodTabs.map((tab) => ({
-            name: canonicalMethodName(tab.name) || String(tab.name || "").trim() || "Unknown Model",
+            name: String(tab.name || "").trim() || "Unknown Model",
             target_price: toNumOrNull(tab.target_price),
             investment_amount: toNumOrNull(tab.investment_amount),
           }))
         : methodBlocks.map((block) => ({
-            name: canonicalMethodName(block.name) || String(block.name || "").trim() || "Unknown Model",
+            name: String(block.name || "").trim() || "Unknown Model",
             target_price: toNumOrNull(block.target_price),
             investment_amount: toNumOrNull(block.investment_amount),
           }));
@@ -196,7 +195,7 @@ export function computeHitRateAggregation(
     }
 
     const hasDreamTeamOutputs = methodTabs.some(
-      (tab) => canonicalMethodName(tab.name).toLowerCase() === "dream team" && Array.isArray(tab.outputs) && tab.outputs.length > 0,
+      (tab) => String(tab.name || "").trim().toLowerCase() === "dream team" && Array.isArray(tab.outputs) && tab.outputs.length > 0,
     );
     if (!hasDreamTeamOutputs) {
       const dreamTeam = Array.isArray(payload.dream_team) ? payload.dream_team : [];
